move config into dedicated file
